# api/swagger.yaml
basePath: /api/v1
definitions:
  account.ConfirmationResend:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  account.PasswordChange:
    properties:
      new:
        type: string
      old:
        type: string
    required:
    - new
    - old
    type: object
  account.PasswordReset:
    properties:
      password:
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  account.Recovery:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  common.StatusMessage:
    properties:
      message:
        type: string
    type: object
  role.ProfileRole:
    properties:
      id:
        type: string
      name:
        maxLength: 100
        type: string
      quota:
        type: integer
    required:
    - name
    type: object
  user.Credentials:
    properties:
      captcha_token:
        type: string
      email:
        maxLength: 255
        type: string
      password:
        type: string
    required:
    - captcha_token
    - email
    - password
    type: object
  user.Profile:
    properties:
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 255
        type: string
      id:
        type: string
      last_name:
        maxLength: 255
        type: string
      role:
        $ref: '#/definitions/role.ProfileRole'
      source:
        maxLength: 100
        type: string
      status:
        maxLength: 100
        type: string
    required:
    - email
    type: object
...
  user.SignupRequest:
    properties:
      captcha_token:
        type: string
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 255
        type: string
      last_name:
        maxLength: 255
        type: string
      password:
        type: string
    required:
    - captcha_token
    - email
    - password
    type: object
...
info:
  contact: {}
  description: Go-micro-SAAS is a web application skeleton to quickly build new SAAS.
  title: Go-micro-SAAS API
  version: "0.1"
paths:
  /account/confirm:
    get:
      consumes:
      - application/json
      description: Confirms the email address of the user
      parameters:
      - description: Token for the email confirmation
        format: uuid
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Email confirmation endpoint
  /account/password/change:
    put:
      consumes:
      - application/json
      description: Resets the password of the logged in user
      parameters:
      - description: The new and old passwords, required to update the password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.PasswordChange'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Reset password endpoint
  /account/password/reset:
    put:
      consumes:
      - application/json
      description: Resets the password of the logged in user
      parameters:
      - description: The token and new password to reset the current set password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.PasswordReset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Reset password endpoint
  /account/recover:
    put:
      consumes:
      - application/json
      description: Send a password reset email to a user
      parameters:
      - description: The email to send the account recovery to
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.Recovery'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Recover account endpoint
  /account/resend:
    put:
      consumes:
      - application/json
      description: Resends email confirmation for an email address.
      parameters:
      - description: The email to send the confirmation to
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.ConfirmationResend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: Resends email confirmation endpoint
  /account/signin:
    post:
      consumes:
      - application/json
      description: Logs in the user, sets up the JWT authorization
      parameters:
      - description: Credentials provided for signing in
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: User sign in endpoint
...
  /account/signup:
    post:
      consumes:
      - application/json
      description: Signs the user up for the application
      parameters:
      - description: User data provided for the signup
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.StatusMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.StatusMessage'
      summary: User signup endpoint
...
swagger: "2.0"